+++ lib/src/controls/inputs/buttons/button.dart
@@ -1,8 +1,17 @@
 import 'package:fluent_ui/fluent_ui.dart';
 
-const kDefaultButtonZFactor = 0.95;
-
+/// A button gives the user a way to trigger an immediate action.
+///
+/// ![Button Example](https://docs.microsoft.com/en-us/windows/apps/design/controls/images/button.png)
+///
+/// See also:
+///
+///   * [ToggleButton], a button that can be on and off.
+///   * [SplitButton], A button with two sides. One side initiates
+///     an action, and the other side opens a menu.
+///   * <https://docs.microsoft.com/en-us/windows/apps/design/controls/buttons>
 class Button extends BaseButton {
+  /// Creates a button.
   const Button({
     Key? key,
     required Widget child,
@@ -24,26 +33,35 @@ class Button extends BaseButton {
   @override
   ButtonStyle defaultStyleOf(BuildContext context) {
     assert(debugCheckHasFluentTheme(context));
-    final theme = FluentTheme.of(context);
+    final ThemeData theme = FluentTheme.of(context);
     return ButtonStyle(
-      cursor: theme.inputMouseCursor,
-      zFactor: ButtonState.resolveWith((states) {
-        if (states.isPressing) return kDefaultButtonZFactor;
+      elevation: ButtonState.resolveWith((states) {
+        if (states.isPressing) return 0.0;
+        return 0.3;
       }),
+      shadowColor: ButtonState.all(theme.shadowColor),
       padding: ButtonState.all(const EdgeInsets.symmetric(
-        horizontal: 12.0,
-        vertical: 8.0,
+        horizontal: 10.0,
+        vertical: 6.0,
       )),
       shape: ButtonState.all(RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(2.0),
+        side: BorderSide(
+          color: theme.disabledColor.withOpacity(0.75),
+          width: 0.1,
+        ),
+        borderRadius: BorderRadius.circular(4.0),
       )),
       backgroundColor: ButtonState.resolveWith((states) {
         return ButtonThemeData.buttonColor(theme.brightness, states);
       }),
       foregroundColor: ButtonState.resolveWith((states) {
         if (states.isDisabled) return theme.disabledColor;
-        return ButtonThemeData.buttonColor(theme.brightness, states)
-            .basedOnLuminance();
+        return ButtonThemeData.buttonColor(theme.brightness, states).basedOnLuminance().toAccentColor()[
+            states.isPressing
+                ? theme.brightness.isLight
+                    ? 'lighter'
+                    : 'dark'
+                : 'normal'];
       }),
     );
   }