+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Date format: DD/MM/YYYY
 
 ## [next] - [##/##/2021]
 
+- Update `ToggleButton` design.
+- Update `Button` design.
 - Update `RadioButton` design:
   - Added support for radio button content (`RadioButton.content`)
 - Update `ContentDialog` design.